Add files via upload
This map was created as part of an interview assessment. The team and I joined a call, and I was given a topic to create a process map and procedure document for. I then had 30 minutes to ask questions and 24 hours after the call ended to create the two deliverables and turn them in. This is the map that I turned in.